apps/interview-rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Pinecone and Vector Store
pinecone-client==3.0.0
langchain==0.1.7
langchain-community==0.0.10
langchain-openai==0.0.6
langsmith==0.1.4

# LangGraph
langgraph==0.0.12

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# LLM and Embeddings
groq==0.4.2
anthropic==0.7.10

# Meeting/Audio Processing
pydub==0.25.1
librosa==0.10.0
numpy==1.26.2

# Real-time
websockets==12.0
aioredis==2.0.1

# Utilities
redis==5.0.1
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
